Provide a docker environment to run config generation cmds.
There are certain pitfalls with installing the Python3
requirements.txt on other distributions. For instance on the most
recent Debian (buster), we see bdist_wheel failures during pip
install, and the installed Go version is too old. It is much more
convenient and reliable to also use docker to generate configs.